pyproject.toml
[project]
name = "optimize-images"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "Pillow",
    "Python-dotenv",
    "tinify",
]
